Add Event Id Interface

+++ src/shipz/util/GameEvent.java
@@ -6,14 +6,14 @@ import java.util.EventObject;
  * Created by Paul on 06.05.2016.
  */
 public class GameEvent extends EventObject {
-    private int _id;
+    private byte _id;
 
-    public GameEvent(Object source, int eventId) {
+    public GameEvent(Object source, byte eventId) {
         super(source);
         _id = eventId;
     }
 
-    public int getId() {
+    public byte getId() {
         return _id;
     }
 }
